% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CQ_test.R
\name{CQ_test}
\alias{CQ_test}
\title{standarized Q test and p-value}
\usage{
CQ_test(x, y)
}
\arguments{
  \item{x}{multivariate sample}

  \item{y}{multivariate sample}
}
\description{
This function calculates Chen and Qin two sample
multivariate test.
}

