% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/q_stat.R, R/test.R
\name{q_stat}
\alias{q_stat}
\title{Q statistic}
\usage{
q_stat(x, y)

q_stat(x, y)
}
\arguments{
  \item{x}{multivariate sample}

  \item{y}{multivariate sample}

  \item{x}{multivariate sample}

  \item{y}{multivariate sample}
}
\description{
This function calculates Chen and Qin statistic two sample
multivariate test.

This function calculates Chen and Qin statistic two sample
multivariate test.
}

