% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tr2_sigma.R
\name{tr2_sigma}
\alias{tr2_sigma}
\title{Trace of square of variance matrix estimator}
\usage{
tr2_sigma(data)
}
\arguments{
  \item{data}{multivariate sample}
}
\description{
Compute Trace of square of variance matrix estimator.
}

