% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tr_sgxsgy.R
\name{tr_sgxsgy}
\alias{tr_sgxsgy}
\title{Trace of product of variance matrices estimator}
\usage{
tr_sgxsgy(x, y)
}
\arguments{
  \item{x}{multivariate sample}

  \item{y}{multivariate sample}
}
\description{
Compute Trace of the product of two variance matrices
estimator.
}

