% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tr_sgxsgy}
\alias{tr_sgxsgy}
\title{Trace of product of variance matrices estimator}
\usage{
tr_sgxsgy(x, y)
}
\arguments{
\item{x}{multivariate sample}

\item{y}{multivariate sample}
}
\description{
Compute Trace of the product of two variance matrices estimator.
}

